Miracle-IT-Career-Academy-Portfolio-Site/backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==3.10.1
Pillow>=10.0.0
requests==2.28.1
razorpay==1.4.2
djangorestframework-simplejwt==5.2.2
PyMySQL==1.1.0
reportlab==4.4.2
google-generativeai==0.3.2
gunicorn
setuptools